refactor: moved calculate_standard_deviation() and calculate_standard_deciation_weighted() to genomic_position.cpp
